+++ src/Models/ChFavorite.php
@@ -3,7 +3,7 @@
 namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
-use sreamtalk\Traits\UUID;
+use StreamTalk\Traits\UUID;
 
 class ChFavorite extends Model
 {