+++ src/StreamTalkMessenger.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace streamtalk;
+namespace StreamTalk;
 
 use App\Models\ChMessage as Message;
 use App\Models\ChFavorite as Favorite;