+++ src/views/pages/app.blade.php
@@ -51,9 +51,9 @@
     <div class="messenger-messagingView">
         {{-- header title [conversation name] amd buttons --}}
         <div class="m-header m-header-messaging">
-            <nav class="streamtalk-d-flex streamtalk-justify-content-between streamtalk-align-items-center">
+            <nav class="StreamTalk-d-flex StreamTalk-justify-content-between StreamTalk-align-items-center">
                 {{-- header back button, avatar and user name --}}
-                <div class="streamtalk-d-flex streamtalk-justify-content-between streamtalk-align-items-center">
+                <div class="StreamTalk-d-flex StreamTalk-justify-content-between StreamTalk-align-items-center">
                     <a href="#" class="show-listView"><i class="fas fa-arrow-left"></i></a>
                     <div class="avatar av-s header-avatar" style="margin: 0px 10px; margin-top: -5px; margin-bottom: -5px;">
                     </div>